requirements.txt
pygame-ce
pyopengl
pyopengl-accelerate
numpy
ctypes